+++ src/entities/index.ts
@@ -11,5 +11,7 @@ export * from "./variables";
 export * from "./webhooks";
 export * from "./releases";
 export * from "./refs";
+export * from "./members";
+export * from "./search";
 
 // All entities now use the registry pattern with embedded handlers